// src/components/common/CSearchForm/types.ts
// 支持的表单类型
export enum COMPONENT_TYPE {
  // 输入框
  INPUT = 'INPUT',
  // 选择框
  SELECT = 'SELECT',
  // 时间
  DATEPICKER = 'DATEPICKER',
  // 开始时间结束时间
  RANGEPIKCER = 'RANGEPIKCER',
  // 自定义组件
  CUSTOM = 'CUSTOM',
  // 下拉框、输入框组合
  SELECTINPUT = 'SELECTINPUT'
}

export enum PLACEHOLDER {
  INPUT = '请输入',
  SELECT = '请选择'
}